requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
scikit-learn>=1.3
Pillow>=10.0
pyyaml>=6.0
gradio>=4.0
